[media] Fix media_build file matching
When I was experimenting with the media_build tree and trying
'make tar DIR=<git repo>' I kept ending up with just one source in my tar
archive. I couldn't for the life of me understand what was going on until
I realized that I had a copy of a media driver source in the top dir of my
git repository. Because the file pattern was not quoted it would expand to
that particular source and match only that one.

It took me a surprisingly long time before I figured this out :-(

Quoting the pattern fixes this.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>