Fix IR setkeycode operations for kernels < 2.6.36
The drivers/input/input.c file on kernels < 2.6.36 first checks if a
keycode exists with getkeycode, before allowing to write a new keycode.
Due to that, on older kernels, rc-core needs to lie to the input layer,
telling that the key exists, otherwise the operation will fail with
-EINVAL.

This fixes "ir-keycode -c" and "ir-keycode -w" parameters.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>