media_build: add a prepare target
Preparing the compability build is something that can't be done in parallel.
This would make the daily build very slow, so as a temporary fix we make a
prepare target that depends on config-compat.h, Makefile.media and links.

So we can do a 'make prepare' first, followed by a make -j<jobs> to actually
build the modules.

This should really be sorted out properly, but we'll save that for a rainy day.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>