Adds a small warning about the needed packages for the script
There are a few dependencies for this script to work. Most are obvious,
but it helps to say what are those. The package names are the ones
found on Fedora rawhide. If other distros have different requirements,
then maybe we should instead put compilation instructions with per-distro
requirements on a separate file or add some automation to call yum/apt-get
and other package management tools if the needed requirements are not found.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>