Temporary workaround for ubuntu firedtv compilation problems
Ubuntu kernels <= 2.6.36 do not provide the ieee1394 headers in the right
place, so firedtv fails to compile.

It shouldn't compile the firedtv ieee1394 stack at all since it was removed
in 2.6.37, but since the firedtv code still contains a ieee1394 source it
suddenly gets compiled again for older kernels.

For now just disable the IEEE1394 part of this driver using config magic.

Once the firedtv driver is patched so that the IEEE1394 code is removed,
this patch can also be reverted.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>