Add a printk messages for media builds
When someone reports an error, we need to know what are the
applied patches. Also, users need to be warned that using the
backported tree is not the same as running a stable kernel.

This patch adds a logic that patches the subsystem core
in order to indicate that an experimental tree is being
used. It also prints the latest applied patches, in order to
help developers when a bug report is filled.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>