rmmod.pl: Add a logic to allow removing audio modules with pulseaudio
Pulseaudio keeps audio devices opened forever. In order to be able to
remove a device, pulseaudio needs to de-allocate the device.

Unfortunately, pulseaudio recognizes alsa drivers as "module"
(an integer number, not related to the device nodename), and only
allows module removal if running with user matches the console owner.

The logic inside rmmod.pl will now take the above into account. So,
it will detect if pulseaudio is running. If it is, it will:

1) list the pulseaudio modules with "pacmd list-sinks"
   and "pacmd list-sources"

2) detect if any of the modules there is provided by a v4l device;

3) If they're provided by a video device, it removes the module with:
	pactl unload-module 26

Even the above logic is not perfect as, due to a pulseaudio libs bug,
pulseaudio can't detect the name of em28xx-alsa driver, as it uses the
same interface as the video node. Similar hacks will may be needed
for other USB devices, like tm6000 and cx231xx.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>