linux/Makefile: Fix the apply_patches logic
The apply_patches logic stores the name of the kernel used during
the last compilation, in order to detect the need of re-apply
the patches, if something changed. However, due to a change at the
logic, the name of the kernel were changed, but the shell script
'dir' var were not updated anymore.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>