Fix build for Debian GNU/Linux testing (wheezy)
Debian removed Digest::SHA1, in favor of Digest::SHA.

According with:
	http://stackoverflow.com/questions/3420720/what-are-the-advantages-of-digestsha-over-digestsha1

Digest::SHA is part of the perl-core and has a better implementation.
I double-checked it on Fedora distros and RHEL6: at least since Fedora 12
and on RHEL6, both packages are there. As Ubuntu generally follows Debian,
the package should also be there.

So, let's just use the Digest::SHA module.

Please, report if your distro comes with a kernel >= 2.6.32 but the ::SHA1
module is not supported there.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>