media_build: disable VIDEO_SMIAPP_PLL, VIDEO_MT9M032 and VIDEO_MT9P031 on old kernels
VIDEO_SMIAPP_PLL can't build on vanilla kernels older than 2.6.34 as it
requires linux/lcm.h; this is not a problem with the Ubuntu 10.04 2.6.32 kernel
as it includes the new lcm.* lib/header files, so the issue was not detected
before. This fixes the error:

media_build/v4l/smiapp-pll.c:26:23: fatal error: linux/lcm.h: No such file or directory

Also, this patch explicitly disables the MT9M032 and MT9P031 drivers that
depends on APTINA_PLL; this drivers try to autoselect the APTINA_PLL dependency
and so they are not enabled by default on old kernels (where this driver is
already blacklisted) but are still manually selectable by the user through
'make menuconfig' or 'make xconfig', breaking compilation.
So it's better to explicitly blacklist this drivers too.

Signed-off-by: Gianluca Gennari <gennarone@gmail.com
Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>