obsolete.txt: Update it to match the recent changes
Mark the media files that moved to some other place as obsolete.

PS.: This is a lazy patch: not all moved files are here as, instead
of carefully analyzing the moves, I just ran:

	$ cd /lib/modules/3.5.3-1.fc17.x86_64/kernel/drivers/media
	$ find . -type f | sed s,./,,|sed s,.ko$,,

and manually removed the stuff that didn't move (radio, rc, media.ko).

Yet, I suspect that this should be enough to get rid of all the old
stuff on other usual x86 distros. Anyway, this is better than nothing.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>