build: improve it to also allow working with a local branch
In the case of local branch, it will try to reduce the amount
of data copied, by just creating a new workdir, if git-new-workdir
is on the path. Otherwise, it will just create a new clone of
a local tree.

Every time build is called using a new-workdir, it will reset
the branch to point to the named branch. So, any changes at
the git workdir under media will be lost.

Of course, git reflog can be used to recover the old patches,
if needed.

Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>