ensure correct install of modules from drivers/{misc,staging}
v4l/scripts/make_mediafile.pl uses the %instdir hash to collate a list
of directories in media_build/linux which contain Makefiles; it uses
this data when constructing a dependency for the Makefile.media target.
It also uses this hash to collate all the individual module names and
generate code which installs the individual modules.
But it gets the installation directory wrong in the case of modules
outside the linux/drivers/media tree.

To fix, do the collation of source locations and conversion into install
locations as separate steps.

Signed-off-by: vincent.mcintyre@gmail.com
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>